size_t -> int

+++ CompetitiveProgramming/SlidingMinimum/test.cpp
@@ -14,7 +14,7 @@ int main() {
     for (int i=0; i<n; ++i) cout << v[i] << (i<n-1 ? " " : "\n");
     cout << "Window Size : " << K << endl;
     
-    vector<size_t> sm = sliding_minimum(v, K);
+    vector<int> sm = sliding_minimum(v, K);
     
     cout << "Index : ";
     for (int i=0; i<n-K+1; ++i) cout << sm[i] << (i<n-K ? " " : "\n");
